requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart    
fastapi-mail
python-dotenv       